requirements.txt
amqp==2.5.2
anyjson==0.3.3
billiard==3.6.1.0
boto3==1.10.30
botocore==1.13.30
celery==4.3.0
certifi==2019.11.28
chardet==3.0.4
cryptography~=2.8
Django==3.0
django-filter==2.2.0
django-timestampable-model==0.1.4
djangorestframework==3.10.3
django-cors-headers~=3.2.0
docutils==0.15.2
drf-jwt~=1.13.3
idna==2.8
importlib-metadata==1.1.0
jmespath==0.9.4
kombu==4.6.6
Markdown==3.1.1
more-itertools==8.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycurl==7.43.0.3 --global-option="--with-openssl"
Pygments==2.5.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
tally==1.0.0
twilio==6.33.1
urllib3==1.25.7
vine==1.3.0
zipp==0.6.0
